+++ lib/xlsxtream/columns.rb
@@ -41,6 +41,13 @@ module Xlsxtream
           # "Using the Calibri font as an example, the maximum digit width of
           #  11 point font size is 7 pixels (at 96 dpi)"
           #
+          # By observation, though, I note that a different spreadsheet-wide
+          # font size selected via the Workbook's ":font => { :size => ... }"
+          # options Hash entry results in Excel, at least, scaling the given
+          # widths in proportion with the requested font size change. We do
+          # not, apparently, need to do that ourselves and run the calculation
+          # based on the reference 11 point -> 7 pixel figure.
+          #
           width_pixels ||= ((((width_chars * 7.0) + 5) / 7) * 256).truncate() / 256.0
 
           xml << %Q{<col min="#{index + 1}" max="#{index + 1}" width="#{width_pixels}" customWidth="1"/>}